Add docstrings and fix typos.
